refactor: add preview png for readme && adjust structure.
