+++ README.md
@@ -1,6 +1,29 @@
 # Weather-app
 
-A lightweight weather forecast application that delivers accurate weather information for any city. To get started:
+A lightweight weather forecast application that delivers accurate weather information for any city.
+
+![Weather App Preview](./src/assets/preview.png)
+
+## Features
+- Real-time weather forecasts
+- City-specific weather information
+- Clean and intuitive interface
+
+## Getting Started
+
+To launch the application:
 ```bash
 npm run dev
 ```
+
+## Project Structure
+- `src/` - Source code directory
+  - `assets/` - Static assets including weather code definitions and preview images
+  - `components/` - React components
+  - `styles/` - CSS stylesheets
+
+## Technologies Used
+- React
+- TypeScript
+- Vite
+- Open-Meteo API
